+++ x/gov/client/cli/cli_proposal_test.go
@@ -56,3 +56,51 @@ func (s IntegrationTestSuite) TestCreateProposalAssignPermission() {
 	err = cmd.ExecuteContext(ctx)
 	s.Require().NoError(err)
 }
+
+func (s IntegrationTestSuite) TestCreateProposalUpsertDataRegistry() {
+	// Query permissions for role Validator
+	val := s.network.Validators[0]
+
+	// We create some random address where we will give perms.
+	addr, err := types3.AccAddressFromBech32("kira1alzyfq30zjsveet87jlg8jxetwqmr0a22c9uz9")
+	s.Require().NoError(err)
+
+	cmd := cli.GetTxProposalUpsertDataRegistry()
+	_, out := testutil.ApplyMockIO(cmd)
+	clientCtx := val.ClientCtx.WithOutput(out).WithOutputFormat("json")
+
+	ctx := context.Background()
+	ctx = context.WithValue(ctx, client.ClientContextKey, &clientCtx)
+
+	cmd.SetArgs([]string{
+		fmt.Sprintf("%s", "theKey"),
+		fmt.Sprintf("%s", "theHash"),
+		fmt.Sprintf("%s", "theReference"),
+		fmt.Sprintf("%s", "theEncoding"),
+		fmt.Sprintf("%d", 12345),
+		fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
+		fmt.Sprintf("--%s=%s", cli2.FlagAddr, addr.String()),
+		fmt.Sprintf("--%s=true", flags.FlagSkipConfirmation),
+		fmt.Sprintf("--%s=%s", flags.FlagBroadcastMode, flags.BroadcastBlock),
+		fmt.Sprintf("--%s=%s", flags.FlagFees, types3.NewCoins(types3.NewCoin(s.cfg.BondDenom, types3.NewInt(100))).String()),
+	})
+
+	err = cmd.ExecuteContext(ctx)
+	s.Require().NoError(err)
+
+	// Vote Proposal
+	out.Reset()
+	cmd = cli.GetTxVoteProposal()
+	cmd.SetArgs([]string{
+		fmt.Sprintf("%d", 2), // Proposal ID
+		fmt.Sprintf("%d", customgovtypes.OptionYes),
+		fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
+		fmt.Sprintf("--%s=true", flags.FlagSkipConfirmation),
+		fmt.Sprintf("--%s=%s", flags.FlagBroadcastMode, flags.BroadcastBlock),
+		fmt.Sprintf("--%s=%s", flags.FlagFees, types3.NewCoins(types3.NewCoin(s.cfg.BondDenom, types3.NewInt(100))).String()),
+	})
+
+	err = cmd.ExecuteContext(ctx)
+	s.Require().NoError(err)
+	fmt.Printf("%s", out.String())
+}